requirements.txt
flask
sqlalchemy
pymysql
pyjwt
